Some Changes

+++ source/Restore.pm
@@ -23,35 +23,42 @@ sub new
     my $class   = ref($invocant) || $invocant;
     my $self = {
         source  => "",
-		sourcename => "",
+        sourcename => "",
         destination   => "",
-		usertime => "",
-		partial => "",
-        verbosity => Verbosity->new(0);
+        usertime => "",
+        partial => "",
+        verbosity => Verbosity->new(0),
     };
-    return bless ($self, $class);
+    bless ($self, $class);
+    $self->{verbosity}->verbose("New","OK");
+    return $self;
 }
 
 sub addSource{
     my ($self,$Source) = @_;
     $self->{source}=$Source;
+    $self->{verbosity}->verbose("New source added: $self->{source}");
 }
 sub addDestination{
-	my ($self, $destination) = @_;
-	$self->{destination} = $destination;
+    my ($self, $Destination) = @_;
+    $self->{destination} = $Destination;
+    $self->{verbosity}->verbose("New destination added: $self->{destination}");
 }
 sub addSourceName{
-	my ($self, $sourcename) = @_;
-	$self->{sourcename} = $sourcename;
+    my ($self, $Sourcename) = @_;
+    $self->{sourcename} = $Sourcename;
+    $self->{verbosity}->verbose("New source name added: $self->{sourcename}");
 }
 sub addUserTime{
-	my ($self, $usertime) = @_;
-	die "Angegeben Zeit entspricht nicht dem Format!" if $self -> {usertime}  =~ m/^\d{4}_\d{2}_\d{2}_\d{2}_\d{2}_\d{2}/;
-	$self->{usertime} = $usertime;
+    my ($self, $Usertime) = @_;
+    die "The given time is not a valid format!" if (!$Usertime  =~ m/^\d{4}_\d{2}_\d{2}_\d{2}_\d{2}_\d{2}/);
+    $self->{usertime} = $Usertime;
+    $self->{verbosity}->verbose("New usertime added: $self->{usertime}");
 }
 sub addPartial{
-	my($self,$partial) = @_;
-	$self->{partial} = $partial;
+    my($self,$Partial) = @_;
+    $self->{partial} = $Partial;
+    $self->{verbosity}->verbose("New partial added: $self->{partial}");
 }
 sub setVerboseLevel{
     my ($self,$level) =@_;
@@ -60,71 +67,81 @@ sub setVerboseLevel{
 
 sub restore_r
 {
-	my $self = shift;
+    my $self = shift;
     $self->{verbosity}->verbose("Start Restore r.\n","OK");
-	$SourceArchiv = Find_source_r
-	(
-		$self -> {source},
-		$self -> {sourcename},
-		$self -> {usertime},
-	);
-	chop($SourceArchiv);
+    $SourceArchiv = Find_source_r
+    (
+    $self,
+    );
+    chop($SourceArchiv);
     $self->{verbosity}->verbose("Find Source Directory: $SourceArchiv.\n","OK");
     $self->{verbosity}->verbose("The given path is not a directory!\n","WARNING");
-	die if(!(-d $SourceArchiv));
-	RestoreDirectory
-	(
-		$SourceArchiv,
-		$self -> {destination},
-		$self -> {sourcename},
-	);
+    die if(!(-d $SourceArchiv));
+    RestoreDirectory
+    (
+    $self,
+    $SourceArchiv,
+    );
 }
 
 sub restore_rp
 {
     my $self = shift;
     $self->{verbosity}->verbose("Start Restore rp.","OK");
-	$SourceArchiv = Find_source_r
-	(
-		$self -> {source},
-		$self -> {sourcename},
-		$self -> {usertime},
-	);
-	chop($SourceArchiv);
-	$FileorArchivSource = Find_source_rp
-	(
-		$SourceArchiv,
-		$self -> {partial},
-	);
+    $SourceArchiv = Find_source_r
+    (
+    $self,
+    );
+    chop($SourceArchiv);
+    my $partial = $self->{partial};
+    $FileorArchivSource = Find_source_rp
+    (
+    $SourceArchiv,
+    $partial,
+    );
     $self->{verbosity}->verbose("The given Subdirectory or File does not exist!","WARNING");
-	die if (!(-f $FileorArchivSource || -d $FileorArchivSource));
-	$FileorArchivDestination = Find_source_rp
-	(
-		$self -> {destination},
-		$self -> {partial},
-	);
+    die if (!(-f $FileorArchivSource || -d $FileorArchivSource));
+    my $destination = $self->{destination};
+    $FileorArchivDestination = Find_source_rp
+    (
+    $destination,
+    $partial,
+    );
+    
+    if(-d $FileorArchivSource)
+    {
+        File::Path::make_path("$self->{destination}/$self->{partial}");
+        $FileorArchivDestination = "$self->{destination}/$self->{partial}";
+    }
+    elsif(-f $FileorArchivSource)
+    {
+        $self->{verbosity}->verbose("There is no file with the given name.\nThe file will be restored in the destination folder!","WARNUNG");
+        $FileorArchivDestination = $self->{destination};
+    }
+    
     $self->{verbosity}->verbose("The given destination path does not exist!\nPlease set a directory or file with the given name!","WARNING");
-	die if(!(-f $FileorArchivDestination||-d $FileorArchivDestination));
-	if(-d $FileorArchivSource){
+    die if(!(-f $FileorArchivDestination||-d $FileorArchivDestination));
+    if(-d $FileorArchivSource){
         $self->{verbosity}->verbose("Find source directory path: $FileorArchivSource.","OK");
         $self->{verbosity}->verbose("Find destination directory path: $FileorArchivDestination.","OK");
-		RestoreSubDirectory
-		(
-			$FileorArchivSource,
-			$FileorArchivDestination,
-			$self->{partial},
-		);
-	}
-	elsif(-f $FileorArchivSource)
-	{
+        RestoreSubDirectory
+        (
+        $FileorArchivSource,
+        $FileorArchivDestination,
+        $self,
+        );
+    }
+    elsif(-f $FileorArchivSource)
+    {
         $self->{verbosity}->verbose("Find source file path: $FileorArchivSource\n.","OK");
         $self->{verbosity}->verbose("Find destination file path: $FileorArchivDestination\n.","OK");
-		RestoreFile
-		(
-			$FileorArchivSource,
-			$FileorArchivDestination,
-		);
-	}
+        RestoreFile
+        (
+        $FileorArchivSource,
+        $FileorArchivDestination,
+        $self,
+        );
+    }
 }
 
 
@@ -134,39 +151,39 @@ sub restore_rp
 ##########################################################################################
 sub Find_source_r
 {
-	my ($self,$source,$sourcename,$Time) = @_;
-	my $FinalSource = "";
-	my $FileTime= "";
-	my $TmpTime="0000_00_00_00_00_00";
-	opendir(my $dir, $source);
-	while(readdir $dir)
-	{
-		if($_ ne "." and $_ ne ".." and $_ ne ".DS_Store")
-		{
-			@tmp = split(/[<>]/, $_);
-			@tmp1 = split(/_/,$tmp[0]);
-			$ArchivName = $tmp1[0];
-			$ArchivTime = $tmp[1];
-			
+    my $self = shift;
+    my $FinalSource = "";
+    my $FileTime= "";
+    my $TmpTime="0000_00_00_00_00_00";
+    opendir(my $dir, $self->{source});
+    while(readdir $dir)
+    {
+        if($_ ne "." and $_ ne ".." and $_ ne ".DS_Store")
+        {
+            @tmp = split(/[<>]/, $_);
+            @tmp1 = split(/_/,$tmp[0]);
+            $ArchivName = $tmp1[0];
+            $ArchivTime = $tmp[1];
+            
             $self->{verbosity}->verbose("Compare times:\n","OK");
             
-			if($ArchivName eq $sourcename)
-			{
-				if(compare_to($ArchivTime) <= compare_to($Time))
-				{
-                    $self->{verbosity}->verbose("Compare $ArchivTime with $Time");
-					if(compare_to($ArchivTime) > compare_to($TmpTime))
-					{
-						$tmpTime = $ArchivTime;
-						$FinalSource = "${source}/${_}";
-						$FinalTime = $ArchivTime;
-					}
-				}
-			}
-		}
-	}
-	closedir($dir);
-	return "$FinalSource\n";
+            if($ArchivName eq $self->{sourcename})
+            {
+                if(compare_to($ArchivTime) <= compare_to($self->{usertime}))
+                {
+                    $self->{verbosity}->verbose("Compare $ArchivTime with $self->{usertime}");
+                    if(compare_to($ArchivTime) > compare_to($TmpTime))
+                    {
+                        $tmpTime = $ArchivTime;
+                        $FinalSource = "$self->{source}/${_}";
+                        $FinalTime = $ArchivTime;
+                    }
+                }
+            }
+        }
+    }
+    closedir($dir);
+    return "$FinalSource\n";
 }
 
 
@@ -176,9 +193,9 @@ sub Find_source_r
 ##########################################################################################
 sub compare_to
 {
-	my $date = shift;
-	my($y,$m,$d,$H,$M,$S) = split(/_/,$date);
-	return "$y$m$d$H$M$S";
+    my $date = shift;
+    my($y,$m,$d,$H,$M,$S) = split(/_/,$date);
+    return "$y$m$d$H$M$S";
 }
 
 
@@ -188,26 +205,22 @@ sub compare_to
 ##########################################################################################
 sub Find_source_rp
 {
-	my ($self,$Directory,$Sourcename) = @_;
-	opendir Dir, $Directory || die "Can`t open Directory!";
-	my @Files = readdir Dir;
-	closedir Dir;
-	foreach (@Files)
-	{
-		if($_ ne "." and $_ ne ".." and $_ ne ".DS_Store")
-		{
-            $self->{verbosity}->verbose("Search in $Directory\n");
-			if(-f "${Directory}/${_}")
-			{
-				return "${Directory}/${_}" if ($_ eq $Sourcename)
-			}
-			elsif(-d "${Directory}/${_}")
-			{
-				return "${Directory}/${_}" if($_ eq $Sourcename);
-				return Find_source_rp("${Directory}/${_}",$Sourcename);
-			}
-		}
-	}
+    my ($Directory,$partial) = @_;
+    opendir Dir, $Directory || die "Can`t open Directory!";
+    my @Files = readdir Dir;
+    closedir Dir;
+    return "${Directory}/${partial}" if(grep(/$partial/,@Files));
+    foreach (@Files)
+    {
+        if($_ ne "." and $_ ne ".." and $_ ne ".DS_Store")
+        {
+            if(-d "${Directory}/${_}")
+            {
+                return "${Directory}/${_}" if($_ eq $partial);
+                return Find_source_rp("${Directory}/${_}",$partial);
+            }
+        }
+    }
 }
 
 
@@ -216,15 +229,16 @@ sub Find_source_rp
 #				Hilfsfunktion um ein ganzes Verzeichnis wieder herzustellen				 #
 ##########################################################################################
 sub RestoreDirectory{
-	my($self,$SourceArchiv,$Destination,$SourceName) = @_;
-	$DestinationName = DestinationArchiv($Destination,$SourceName);
-    $self->{verbosity}->verbose("Remove ${Destination}/${DestinationName}\n","OK");
-    File::Path::remove_tree("${Destination}/${DestinationName}");
-    $self->{verbosity}->verbose("Make ${Destination}/${SourceName}_<${FinalTime}>\n","OK");
-	File::Path::make_path("${Destination}/${SourceName}_<${FinalTime}>");
-    $self->{verbosity}->verbose("Copy from $SourceArchiv to ${Destination}/${SourceName}_<${FinalTime}>","OK");
-	File::Copy::Recursive::dircopy($SourceArchiv,"${Destination}/${SourceName}_<${FinalTime}>");
+    my($self,$SourceArchiv) = @_;
+    $DestinationName = DestinationArchiv($self->{destination},$self->{sourcename});
+    $self->{verbosity}->verbose("Remove $self->{destination}/${DestinationName}\n","OK");
+    File::Path::remove_tree("$self->{destination}/${DestinationName}") if(-d "$self->{destination}/${DestinationName}");
+    $self->{verbosity}->verbose("Make $self->{destination}/$self->{sourcename}}_<${FinalTime}>\n","OK");
+    File::Path::make_path("$self->{destination}/$self->{sourcename}_<${FinalTime}>");
+    $self->{verbosity}->verbose("Copy from $SourceArchiv to $self->{destination}}/$self->{sourcename}_<${FinalTime}>","OK");
+    File::Copy::Recursive::dircopy($SourceArchiv,"$self->{destination}/$self->{sourcename}_<${FinalTime}>");
 }
+
 
 
 ##########################################################################################
@@ -232,16 +246,16 @@ sub RestoreDirectory{
 #				Hilfsfunktion um die genau Adresse des Zielverzeichnis zu erhalten		 #
 ##########################################################################################
 sub DestinationArchiv{
-	my ($Destination,$SourceName) = @_;
-	opendir DIR,$Destination;
-	my @DestinationFiles = readdir DIR;
-	closedir DIR;
-	foreach (@DestinationFiles)
-	{
-		if ($_ =~ m/${SourceName}*/){
-			return $_;
-		}
-	}
+    my ($Destination,$SourceName) = @_;
+    opendir DIR,$Destination;
+    my @DestinationFiles = readdir DIR;
+    closedir DIR;
+    foreach (@DestinationFiles)
+    {
+        if ($_ =~ m/${SourceName}*/){
+            return $_;
+        }
+    }
 }
 
 
@@ -250,14 +264,14 @@ sub DestinationArchiv{
 #				Hilfsfunktion um ein Unterverzeichnis wieder herzustellen				 #
 ##########################################################################################
 sub RestoreSubDirectory{
-	my($self,$Source,$Destination,$SourceName)=@_;
-	$FatherDir = DirectoryUp($Destination);
-    $self->{verbosity}->verbose("Remove $Destination\n","OK");
-	File::Path::remove_tree($Destination) if ($_Verbose eq "0");
-    $self->{verbosity}->verbose("Make Dir: ${FatherDir}/${SourceName}\n","OK");
-	File::Path::make_path("${FatherDir}/${SourceName}");
-    $self->{verbosity}->verbose("Copy Dir from $Source to ${FatherDir}/${SourceName}\n","OK");
-	File::Copy::Recursive::dircopy($Source, "${FatherDir}/${SourceName}");
+    my($Source,$Destination,$self)=@_;
+    $FatherDir = DirectoryUp($Destination);
+    $self->{verbosity}->verbose("Remove $Destination\n","OK")if(-d $Destination);
+    File::Path::remove_tree($Destination) if(-d $Destination);
+    $self->{verbosity}->verbose("Make Dir: ${FatherDir}/$self->{partial}\n","OK");
+    File::Path::make_path("${FatherDir}/$self->{partial}");
+    $self->{verbosity}->verbose("Copy Dir from $Source to ${FatherDir}/$self->{partial}\n","OK");
+    File::Copy::Recursive::dircopy($Source, "${FatherDir}/$self->{partial}");
 }
 
 
@@ -266,12 +280,20 @@ sub RestoreSubDirectory{
 #				Hilfsfunktion um eine Datei wieder herzustellen							 #
 ##########################################################################################
 sub RestoreFile{
-	my($self,$Source,$Destination)=@_;
-	$FatherDir = DirectoryUp($Destination);
-    $self->{verbosity}->verbose("Delete Destinationfile: $Destination.\n");
-	unlink($Destination);
-    $self->{verbosity}->verbose("Copy File to $FatherDir\n");
-	File::Copy::copy($Source,$FatherDir);
+    my($Source,$Destination,$self)=@_;
+    if(-f $Destination)
+    {
+        my $FatherDir = DirectoryUp($Destination) if(-f $Destination);
+        $self->{verbosity}->verbose("Delete Destinationfile: $Destination.\n");
+        unlink($Destination);
+        $self->{verbosity}->verbose("Copy File to $FatherDir\n");
+        File::Copy::copy($Source,$FatherDir);
+    }
+    elsif(-d $Destination)
+    {
+        $self->{verbosity}->verbose("Copy file to $Destination","OK");
+        File::Copy::copy($Source,$Destination);
+    }
 }
 
 
@@ -280,19 +302,19 @@ sub RestoreFile{
 #				Hilfsfunktion um ein Verzeichnis hoch zu wechseln						 #
 ##########################################################################################
 sub DirectoryUp{
-	@tmp = split(/\//,shift);
-	$Return ="";
-	for(my $i = 1;$i < scalar(@tmp)-1; $i++)
-	{
-		$Return .= "/${tmp[$i]}";
-	}
-	return $Return;
+    @tmp = split(/\//,shift);
+    $Return ="";
+    for(my $i = 1;$i < scalar(@tmp)-1; $i++)
+    {
+        $Return .= "/${tmp[$i]}";
+    }
+    return $Return;
 }
 
 
 sub DESTROY{
     my $self = shift;
-	$FinalTime ="";
+    $FinalTime ="";
     $self->{verbosity}->verbose("Destroy");
 };
 1;
